css 수정

+++ src/main/java/com/project/member/dto/MemberDTO.java
@@ -10,7 +10,7 @@ import java.util.Date;
 @Setter
 @ToString
 public class MemberDTO {
-    private Long m_idx;
+    private int m_idx;
     private String m_id;
     private String m_pw;
     private String m_email;
